assignment

+++ exp/06-eigen/build.sh
@@ -2,7 +2,9 @@
 mkdir build
 cd build
 
-rm -rf *    # delete all contents in this "build" folder
+if [[ $1 == "rm" ]]; then
+	rm -rf *
+fi
 
 cmake ..
 make -j4
